requirements.txt
asgiref==3.7.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==41.0.1
defusedxml==0.7.1
dj-rest-auth==4.0.1
Django==4.2.1
django-allauth==0.54.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
idna==3.4
joblib==1.2.0
numpy==1.24.3
oauthlib==3.2.2
pandas==2.0.2
Pillow==9.5.0
psycopg==3.1.9
psycopg-binary==3.1.9
pycparser==2.21
PyJWT==2.7.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sklearn==0.0.post5
sqlparse==0.4.4
threadpoolctl==3.1.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==2.0.2
